Made code display and tabs server components.
