+++ src/app/LandingSegment.tsx
